# ARROW-11502: [C++] Optimize Arrow ByteStreamSplitDecode with Neon