Improve ByteStreamSplitDecode performance for float_1k/4k/32k/64k:
1k:  21.6536G/s -> 27.7227G/s
4K:  22.1561G/s -> 28.2019G/s
32k: 21.9854G/s -> 25.8939G/s
64k: 22.0192G/s -> 26.1399G/s
No obvious performance difference for Double_1k/4k/32k/64k.