@guyuqi 2021-02-23
From discussion on dev mailing-list,  the soft consensus among us is that adopting [xsimd](https://github.com/xtensor-stack/xsimd) as the SIMD wrapper library for Arrow. 

I'm going to contribute the missing functions to xsimd first.
So convert to draft.