@pitrou 2021-02-23
I opened https://issues.apache.org/jira/browse/ARROW-11744 to track xsimd vendoring.