@guyuqi 2021-05-12
Refactor Decoding-Bytestream implementation with XSIMD
Improve ByteStreamSplitDecode performance by leveraging simd library:xsimd.

The tests of Byte-streaming-decode(parquet-internals-test)are passed on x86/Arm64:
```
[----------] 4 tests from TestByteStreamSplitEncoding/0, where TypeParam = parquet::PhysicalType<(parquet::Type::type)4>
[ RUN      ] TestByteStreamSplitEncoding/0.BasicRoundTrip
[       OK ] TestByteStreamSplitEncoding/0.BasicRoundTrip (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/0.RoundTripSingleElement
[       OK ] TestByteStreamSplitEncoding/0.RoundTripSingleElement (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/0.CheckOnlyDecode
[       OK ] TestByteStreamSplitEncoding/0.CheckOnlyDecode (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/0.CheckOnlyEncode
[       OK ] TestByteStreamSplitEncoding/0.CheckOnlyEncode (0 ms)
[----------] 4 tests from TestByteStreamSplitEncoding/0 (0 ms total)

[----------] 4 tests from TestByteStreamSplitEncoding/1, where TypeParam = parquet::PhysicalType<(parquet::Type::type)5>
[ RUN      ] TestByteStreamSplitEncoding/1.BasicRoundTrip
[       OK ] TestByteStreamSplitEncoding/1.BasicRoundTrip (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/1.RoundTripSingleElement
[       OK ] TestByteStreamSplitEncoding/1.RoundTripSingleElement (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/1.CheckOnlyDecode
[       OK ] TestByteStreamSplitEncoding/1.CheckOnlyDecode (0 ms)
[ RUN      ] TestByteStreamSplitEncoding/1.CheckOnlyEncode
[       OK ] TestByteStreamSplitEncoding/1.CheckOnlyEncode (0 ms)
[----------] 4 tests from TestByteStreamSplitEncoding/1 (0 ms total)

[----------] 1 test from ByteStreamSplitEncodeDecode
[ RUN      ] ByteStreamSplitEncodeDecode.InvalidDataTypes
[       OK ] ByteStreamSplitEncodeDecode.InvalidDataTypes (0 ms)
[----------] 1 test from ByteStreamSplitEncodeDecode (0 ms total)
```